Moved Position to its own file
